feat: Add Redux Toolkit and React-Redux to the project
This commit adds "@reduxjs/toolkit": "^1.9.5" and "react-redux": "^8.1.0" as dependencies in package.json file, which are used for state management in the application.

It also includes changes to App.tsx, GameBoard.scss, GameBoard.tsx, ScoreComponent.scss, ScoreComponent.tsx files to use Redux store with React-Redux hooks.

Additionally, it introduces new components such as BallComponent.tsx, BoardWithNextBallsComponent.scss and BoardWithNextBallsComponent.tsx that are used for rendering game board elements.

Finally, BFSAlgorithm.helper.ts is added to implement a helper function for Breadth First Search algorithm that is used in the game logic.